NRA-0004: publish v1.3.1 bug fix

+++ nodes/aws/sqs-read.js
@@ -72,7 +72,7 @@ module.exports = function(RED) {
                 send([null, error]);
             }
 
-            if (messageListToDelete.length) {
+            if (node.deleteAfterRead && messageListToDelete.length) {
                 await deleteMessage(client, sqsUrl, messageListToDelete);
             }
